Initial board rendering, loading files, creation and definition of classes
In this commit, the basic features of the program were implemented, although not in a full way.
The program is capable of loading the pieces, generate a board and insert a piece, through the main function. This creates the necessary steps to allow to proceed to implementing the game logic, loop, move validation and game states.

The main challenge till now was getting the pieces to render properly and with quality in the screen. One discover that I can say is relatively to the Macbook Pro Retina Screen. Enabling the High DPI mode fixed the issue of low resolution, but introduced the fact that the code should correct the difference between the logical position and the actual position of the screen. This was achieved using the SDL_SetScale