Implementation of the FEN Parser, Inheritance and Polymorphism Mechanism
In this commit, the FEN Parser works to place the pieces into an array of Piece * and then use polymorphism concepts in order to store the respective pieces. Of course, it might not be the most efficient but the goal is to test and implement diverse C++ concepts in order to learn new concepts. The pieces can be rendered through some lines of code in the main, which will be refined and transferred to the respective class that is responsible for rendering. Also, each class got most of their methods and attributes, in order to setup the grounds to create each virtual function.

I can say that it was a very interesting experience to reach this point