Code Refactoring, Early Implementation of Move Generation and Move Highlighting
In this commit, some code refactoring and grouping were made, to facilitate the appropriate render of diverse renders in the correct sequence. After that, the early implementation of the virtual functions of the pieces were created and the respective graphical function which allow the user to visualize the possible moves. The possible moves are generated for the Knight (without wrap check for now), Bishop and Rook (with wrap check). I expect to finish all piece movement today and work more on the interface to allow actually the pieces to be moved and the underlying attributes to be changed accordingly.