Full Move Generation and Move/Capture Piece Logic
In this new version, the functionality of each piece generating their moves was fully implemented (although still special moves are not implemented and managed). So the piece is able to get have all the generated moves, display moves and captures. Now, the early functionality of allowing pieces to move and capture was implemented.

A redesign in the way that Pieces and Board interact is planned, to follow a more OOP approach: the Pieces will own a Board * to the board there are inserted, so it is not necessary to pass the reference to the owner board.