Implementation of a Game Class, Piece Capture and Animations
In this new commit, the Game Class was implemented, which is responsible for holding and associate the logical and graphical class together, and run the game loop, perform event handling (such as move, drag, and capture) and handling turns (white or black). The logic for capturing pieces and moving pieces was implemented, together with the respective animation. The game is playable, now reaching the stage of being pseudolegal. The next steps will be implementing new details and rules, such as castling, en-passant, add sounds effect and maybe letters to the side of the board using the respective SDL libraries. Overall, the project is in good shape :D