First WORKING Implementation of Dear imgui
This commit contains the first working implementation of the dear imgui API for a GUI. One important note, don't return the Graphics without it being a reference.

Some refactors were made in the Graphics and in the Game class, especially to allow rendering the board every 120fps, not ondemand or based on event triggering. For now, the imgui is on main but it will be moved to a ChessGUI class