Creation of ChessGUI class using imgui as GUI API
The class ChessGUI was created to accomodate all the related functions, methods and attributes to allow for a seamless integration to the existing board graphical interface. The relevant methods were added to allow different functionalities such as loading a FEN, resetting board, hiding markings and adding some flexibility for changing the colors of the board. Also, most of the basis for a implementation of an Chess AI Engine are implemented, which will be a possible project for the Summer.